% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_selection.R
\name{aggregateScoreChangePerGene}
\alias{aggregateScoreChangePerGene}
\title{Aggregate Score Change per Gene}
\usage{
aggregateScoreChangePerGene(
  d.scores,
  annotation,
  marker.type,
  cell.type,
  target.type = NULL,
  balance.cell.types = T,
  self.mult = 1
)
}
\description{
Aggregates score chage over all type. Either average (`balance.cell.types=T`)
or sum (`balance.cell.types=F`) it per cell type and returns sum of signed scores
}
