% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alldataset.R
\docType{data}
\name{alldataset}
\alias{alldataset}
\title{A list of data set after split reference data into train and test data set}
\format{
A list with trainResponseData, testResponseData, trainExplanatoryData and testExplanatoryData
}
\usage{
data(alldataset)
}
\description{
A list containing 4 data set of train and test response data and explanatory data
}
\author{
Junhui Li \email{junhui.li11@umassmed.edu}
}
\keyword{data}
