% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{arrangePlots}
\alias{arrangePlots}
\title{Arrange Plots}
\usage{
arrangePlots(plot.list, build.panel, n.col = NULL, n.row = NULL)
}
\arguments{
\item{plot.list}{list of plots}

\item{build.panel}{combine individual plots to a single panel}

\item{n.col}{number of columns in the panel}

\item{n.row}{number of rows in the panel}
}
\value{
panel of plots if `build.panel==T` or `plot.list` otherwise
}
\description{
Arrange Plots
}
