% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{assignCellsByScores}
\alias{assignCellsByScores}
\title{Assign Cells By Scores}
\usage{
assignCellsByScores(graph, clf.data, score.info = NULL,
  clusters = NULL, verbose = 0, uncertainty.thresholds = c(coverage =
  0.5, negative = 0.5, positive = 0.75), ...)
}
\arguments{
\item{graph}{cell graph from Seurat, Pagoda2 or some other tool. Can be either in igraph or adjacency matrix format.
Use `graph=NULL` to skip graph diffusion step and get raw score annotation (useful when debug marker genes).}

\item{clf.data}{classification data from `getClassificationData`}

\item{score.info}{cell type scores from `getMarkerScoreInfo` function. Re-estimated if NULL}

\item{clusters}{vector with cluster labels named by cell ids. Used to expand annotation on these clusters.}

\item{verbose}{verbosity level (from 0 to 2)}

\item{...}{Arguments passed on to \code{diffuseScorePerType}
\describe{
  \item{}{}
}}
}
\value{
list with parameters:\itemize{
  \item{annotation: annotation per level}
  \item{scores: assignment scores per level}
  \item{annotation.filt: the same as annotation, but cells, which don't pass QC are assigned to NA class}
}
}
\description{
Assign cell types for each cell based on type scores. Optionally uses `clusters` to expand annotation.
}
\examples{
  clf_data <- getClassificationData(cm, marker_path)
  ann_by_level <- assignCellsByScores(graph, clf_data, clusters=clusters)

}
