% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildSVMclassifier.R
\name{buildSVMclassifier}
\alias{buildSVMclassifier}
\title{build SVM classifier}
\usage{
buildSVMclassifier(
  x,
  y,
  scale = TRUE,
  type = c("C-classification"),
  kernel = c("radial", "linear", "polynomial", "sigmoid")
)
}
\arguments{
\item{x}{explanatory data set.}

\item{y}{a response vector with one label for each row/component of x. And this should be a factor vector.}

\item{scale}{A logical vector indicating the variables to be scaled. If scale is of length 1, the value is recycled as many times as needed. Per default, data are scaled internally (both x and y variables) to zero mean and unit variance. The center and scale values are returned and used for later predictions.}

\item{type}{svm can be used as a classification machine in this function. The default setting for type is C-classification.Valid options are:
C-classification}

\item{kernel}{the kernel used in training and predicting. see kernel of svm for detail message.}
}
\description{
build SVM classifier as a classification machine
}
\examples{

data(uniqueFeatureGene)
data(alldataset)
trainNorExp <- normalizeSCExp(alldataset$trainExplanatoryData,dThresh = 0.25)
x=t(as.matrix(trainNorExp[uniqueFeatureGene,]))
y=as.factor(alldataset$trainResponseData$cellType)
svmClassifier <- buildSVMclassifier(x,y)

}
