% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{classificationTreeToDf}
\alias{classificationTreeToDf}
\title{Classification Tree to DataFrame}
\usage{
classificationTreeToDf(classification.tree)
}
\arguments{
\item{classification.tree}{cell type hierarchy represented by graph. Part of `clf_data` object from `getClassificationData`}
}
\description{
Classification Tree to DataFrame
}
