% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markup_processing.R
\name{createClassificationTree}
\alias{createClassificationTree}
\title{Create Classification Tree}
\usage{
createClassificationTree(marker.list)
}
\arguments{
\item{marker.list}{list of markers per cell type. Can be obtained with `parseMarkerFile`}
}
\value{
igraph graph with the type hierarhcy
}
\description{
creates graph object for cell type hierarchy
}
\examples{
  markers <- parseMarkerFile(marker_path)
  clf_tree <- createClassificationTree(markers)

}
