% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchies.R
\name{deriveHierarchy}
\alias{deriveHierarchy}
\title{Derive Hierarchy}
\usage{
deriveHierarchy(feature.matrix, annotation, dist.method = "cor", max.depth = 2)
}
\arguments{
\item{feature.matrix}{matrix where rows represent cells and columns represent either genes or some embedded space (e.g. PCA)}

\item{annotation}{vector with cell type label per cell}

\item{dist.method}{method for pairwise distance estimation. Either "cor" for correlation distance or any method supported by `dist` function}

\item{max.depth}{maximal depth of the hierarchy}
}
\value{
list with cell type hierarchy
}
\description{
derive hierarchy from the data using hclust
}
\examples{
  hierarchy <- deriveHierarchy(pca_mtx, annotation, max.depth=3)
  clf_tree <- hierarchyToClassificationTree(hierarchy)
  plotTypeHierarchy(clf_tree)

}
