% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{diffuseGraph}
\alias{diffuseGraph}
\title{Diffuse Graph}
\usage{
diffuseGraph(
  graph,
  scores,
  fading = 10,
  fading.const = 0.5,
  score.fixing.threshold = 0.8,
  verbose = FALSE,
  max.iters = 1000,
  tol = 0.001,
  smoothing.method = c("heat.filter", "knn"),
  smooth.strength = 30
)
}
\arguments{
\item{graph}{graph to diffuse on}

\item{scores}{table of scores}

\item{fading}{fading level for graph diffusion}

\item{fading.const}{constant in exponent for graph diffusion}

\item{score.fixing.threshold}{threshold for a label to be considered certain. Such labels can't be changed during diffusion.}

\item{verbose}{print progress bar}

\item{tol}{tolerance for diffusion stopping}

\item{smoothing.method}{method to use for signal smoothing: `heat.filter` corresponds to smooting using graph filters (see \link[sccore:smoothSignalOnGraph]{smoothSignalOnGraph}),
while `knn` corresponds to kNN diffusion (see \link[sccore:smooth_count_matrix]{smooth_count_matrix}).}

\item{smooth.strength}{smoothing strength (`beta`) for `smoothing.method='heat.filter'`}
}
\description{
Run diffusion on graph
}
