% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{diffuseScorePerType}
\alias{diffuseScorePerType}
\title{Diffuse Score per Type}
\usage{
diffuseScorePerType(
  scores.per.type,
  graph,
  parents,
  cbs.per.type,
  verbose,
  n.cores = 1,
  ...
)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=diffuseGraph]{diffuseGraph}}
  \describe{
    \item{\code{fading}}{fading level for graph diffusion}
    \item{\code{fading.const}}{constant in exponent for graph diffusion}
    \item{\code{score.fixing.threshold}}{threshold for a label to be considered certain. Such labels can't be changed during diffusion.}
    \item{\code{verbose}}{print progress bar}
    \item{\code{tol}}{tolerance for diffusion stopping}
    \item{\code{smoothing.method}}{method to use for signal smoothing: `heat.filter` corresponds to smooting using graph filters (see \link[sccore:smoothSignalOnGraph]{smoothSignalOnGraph}),
while `knn` corresponds to kNN diffusion (see \link[sccore:smooth_count_matrix]{smooth_count_matrix}).}
    \item{\code{beta}}{smoothing strength for `smoothing.method='heat.filter'`}
  }}
}
\description{
Diffuse Score per Type
}
