% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{diffuseSubgraph}
\alias{diffuseSubgraph}
\title{Run diffusion on subgraph for the specified subtype}
\usage{
diffuseSubgraph(graph, annotation, subtype, scores, fading = 10,
  fading.const = 0.5, clusters = NULL, verbose = FALSE,
  max.iters = 1000, tol = 0.001)
}
\arguments{
\item{fading}{fading level for graph diffusion}

\item{fading.const}{constant in exponent for graph diffusion}

\item{verbose}{print progress bar}

\item{tol}{tolerance for diffusion stopping}
}
\description{
Run diffusion on subgraph for the specified subtype
}
