% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalMetrics.R
\docType{data}
\name{evalMetrics}
\alias{evalMetrics}
\title{A list of data set outputted from evaluateClassifier function}
\format{
A list with dataset named "multiLogLoss", "confusionMatrix" and "rocObjList"
}
\usage{
data(evalMetrics)
}
\description{
A list containing 3 data set of "multiLogLoss", "confusionMatrix" and "rocObjList"
}
\author{
Junhui Li \email{junhui.li11@umassmed.edu}
}
\keyword{data}
