% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateClassifier.R
\name{evaluateClassifier}
\alias{evaluateClassifier}
\title{evaluate the classifier}
\usage{
evaluateClassifier(predY, queryY)
}
\arguments{
\item{predY}{output of predict result from predict function}

\item{queryY}{A vector of response variable named with independent variables}
}
\description{
Computing Classification Evaluation Metrics
}
\examples{

data(uniqueFeatureGene)
data(alldataset)
trainNorExp <- normalizeSCExp(alldataset$trainExplanatoryData,dThresh = 0.25)
x=t(as.matrix(trainNorExp[uniqueFeatureGene,]))
y=as.factor(alldataset$trainResponseData$cellType)
svmClassifier <- buildSVMclassifier(x,y)
testNorExp <- normalizeSCExp(alldataset$testExplanatoryData,dThresh = 0.25)
testX=t(as.matrix(testNorExp[uniqueFeatureGene,]))
yPred <- predictWithclassifier(svmClassifier,testX)
yActual <- as.factor(alldataset$testResponseData$cellType)
names(yActual) <- alldataset$testResponseData$cell
evaluatedResult <- evaluateClassifier(yPred,yActual)

}
