% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{extractCellTypeProbs}
\alias{extractCellTypeProbs}
\title{Extract Cell Type Probabilities}
\usage{
extractCellTypeProbs(annotation.scores, clf.tree, ann.level = NULL)
}
\value{
data.frame of annotation probabilities with cell by rows and cell types by columns
}
\description{
Re-normalizes scores to estimate full probability for each cell to belong to a specific cell type
}
