% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{extractMarkersFromSubtypes}
\alias{extractMarkersFromSubtypes}
\title{Extract Markers From Subtypes}
\usage{
extractMarkersFromSubtypes(parent.type = "root", clf.data = NULL,
  clf.tree = NULL, marker.list = NULL, count.matrix = NULL,
  max.depth = NULL, drop.missing = T, marker.type = c("expressed",
  "not_expressed"))
}
\arguments{
\item{parent.type}{cell type for which the markers of the subtypes must be plotted}
}
\description{
Extract Markers From Subtypes
}
