% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterSCGene.R
\name{filterSCGene}
\alias{filterSCGene}
\title{filter single cell genes}
\usage{
filterSCGene(
  expDat,
  mu = 2,
  alpha1 = 0.1,
  alpha2 = 0.01,
  threshold = 0,
  pseudo = FALSE
)
}
\arguments{
\item{expDat}{single cell expression dataset}

\item{mu}{the threshold of average expression of genes passing}

\item{alpha1}{the proportion of cells in which a gene must be considered detected}

\item{alpha2}{the proportion of cells for genes that have higher expression level}

\item{threshold}{threshold of gene expression detection}

\item{pseudo}{pseudo of calculation of the proportion of cells}
}
\description{
filter single cell genes based on the idea that reliably detected genes will either be detected in many cells, or highly expressed in a small cels of cells (or both)
}
\examples{

data(alldataset)
trainNorExp <- normalizeSCExp(alldataset$trainExplanatoryData,dThresh = 0.25)
filterGene <- filterSCGene(trainNorExp,mu = 2,alpha1 = 0.1,alpha2 = 0.01,threshold=0.25)

}
