% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findFeatureGene.R
\name{findFeatureGene}
\alias{findFeatureGene}
\title{find feature gene}
\usage{
findFeatureGene(
  responseData,
  explanatoryData,
  filteredGene,
  topX = 50,
  reverse = TRUE
)
}
\arguments{
\item{responseData}{response vector}

\item{explanatoryData}{explanatory dataset}

\item{filteredGene}{output of filterSCGene}

\item{topX}{a numeric vector to select the top X genes}

\item{reverse}{select the bottom X genes, logistic value, defaul is TRUE.}
}
\description{
find feature gene based on least square method
}
\examples{

data(alldataset)
trainNorExp <- normalizeSCExp(alldataset$trainExplanatoryData,dThresh = 0.25)
filterGene <- filterSCGene(trainNorExp,mu = 2,alpha1 = 0.1,alpha2 = 0.01,threshold=0.25)
responseData <- alldataset$trainResponseData$cellType
uniqueFeatureGene <- findFeatureGene(responseData,trainNorExp,filteredGene=filterGene,
topX = 20,reverse=TRUE)

}
