% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{getCellTypeScoreInfo}
\alias{getCellTypeScoreInfo}
\title{Get Cell Type Score Info}
\usage{
getCellTypeScoreInfo(markers, tf.idf, aggr = T)
}
\arguments{
\item{markers}{element of marker list. List with fields `expressed` and `not_expressed`}

\item{tf.idf}{TF-IDF normalized matrix. Can be obtained with `normalizeTfIdfWithFeatures`}

\item{aggr}{if scores must be aggregated for the whole cell type or returned for each marker separately}
}
\value{
list with score info:\itemize{
  \item{scores.raw: scores from positive markers}
  \item{mult: score multiplier, estimated with negative markers}
  \item{max.positive: maximal expression of positive markers. Used for estimation of negative scores}
  \item{scores: final scores. Equal to `scores * score.mult`}
}
}
\description{
estimate info, neccessary for scoring of cells by cell types for the specified cell type
}
