% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getMarkerScoresPerCellType}
\alias{getMarkerScoresPerCellType}
\title{Return initial scores of each cell type for each cell}
\usage{
getMarkerScoresPerCellType(clf.data, score.info = NULL, aggr = T)
}
\arguments{
\item{clf.data}{classification data from `getClassificationData`}

\item{score.info}{pre-calculated score info from `getMarkerScoreInfo`}

\item{aggr}{should individual gene scores be aggregated per cell type? If `FALSE`,
returns list of data.frames, showing scores of each gene for each cell.
Useful for debugging list of markers.}
}
\value{
data.frame with rows corresponding to cells and columns corresponding to cell types.
  Values are cell type scores, normalized per level of hierarchy
}
\description{
Return initial scores of each cell type for each cell
}
