% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_selection.R
\name{getTopNegativeGenes}
\alias{getTopNegativeGenes}
\title{Get Top Negative Genes}
\usage{
getTopNegativeGenes(
  pos.gene,
  cell.type,
  cm.norm,
  annotation,
  cur.neg.genes,
  s.info,
  pos.score.changes,
  n.neg.genes,
  score.change.threshold
)
}
\arguments{
\item{pos.gene}{positive marker for which new negative markers must be found}

\item{cell.type}{all markers are relative to this cell type}

\item{cm.norm}{tf-idf normalized count matrix}

\item{annotation}{cell type annotation}

\item{cur.neg.genes}{negative markers candidates}

\item{s.info}{score info}

\item{pos.score.changes}{score changes per cell from the positive marker candidates. Only `pos.score.changes[,pos.gene]` is used}

\item{n.neg.genes}{number of new negative markers to select}

\item{score.change.threshold}{???}
}
\description{
Get Top Negative Genes
}
