% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchies.R
\name{hierarchyToClassificationTree}
\alias{hierarchyToClassificationTree}
\title{Hierarchy to Classification Tree}
\usage{
hierarchyToClassificationTree(hierarchy)
}
\arguments{
\item{hierarchy}{list of cell types, where each element represent a cell type. If cell type has some subtypes it's represented as another list, otherwise it's just a string}
}
\value{
igraph graph with the type hierarhcy
}
\description{
Convert list with cell type hierarchy to classification tree in igraph format
}
\examples{
  hierarchy <- list(
    Alveolar=c("AT1 Cell", "AT2 Cell"),
    B=c("B Cell", "Ig-producing B cell"),
    NK_T=list(`T`=c("Dividing T cells", "T Cell_Cd8b1 high", "Nuocyte"), "NK Cell"),
    "Endothelial"
  )

  clf_tree <- hierarchyToClassificationTree(hierarchy)

}
