% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garnett_wrappers.R
\name{markerListToMarkup}
\alias{markerListToMarkup}
\title{Convert marker list to the markup language and optionally save it to a file}
\usage{
markerListToMarkup(marker.list, file = "", group.by.parent = T)
}
\arguments{
\item{marker.list}{list of markers}

\item{file}{file to save. If empty string, returns markup instead of saving}

\item{group.by.parent}{group cell types in the markup by the parent type}
}
\description{
Convert marker list to the markup language and optionally save it to a file
}
