% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mergeAnnotationByLevels}
\alias{mergeAnnotationByLevels}
\title{Merge Annotation By Levels}
\usage{
mergeAnnotationByLevels(annotation, classification.tree)
}
\arguments{
\item{annotation}{annotation for high-resolution. Cell type names must correspond to nodes in the `classification.tree`}

\item{classification.tree}{cell type hierarchy represented by graph. Part of `clf_data` object from `getClassificationData`}
}
\value{
list of annotations where each element correspond to some hierarchy level
}
\description{
merge provided annotation using `classification.tree` to get annotations for different levels of hierarchy
}
\examples{
  ann_by_level <- mergeAnnotationByLevels(annotation, clf_data$classification.tree)

}
