% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mergeAnnotationToLevel}
\alias{mergeAnnotationToLevel}
\title{Merge Annotation to Level}
\usage{
mergeAnnotationToLevel(level, annotation, classification.tree)
}
\arguments{
\item{annotation}{annotation for high-resolution. Cell type names must correspond to nodes in the `classification.tree`}
}
\description{
Merge Annotation to Level
}
