% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mergeScoreInfos}
\alias{mergeScoreInfos}
\title{Merge Score Infos}
\usage{
mergeScoreInfos(score.infos, verbose = F)
}
\arguments{
\item{score.infos}{scoring info per cell per cell type}

\item{verbose}{show progress bar}
}
\value{
List of score info for each cell type. See `CellAnnotatoR:::getCellTypeScoreInfo` for more info
}
\description{
merge score information from multiple datasets
}
\examples{
  clf_datas <- lapply(cms, getClassificationData, marker_path)
  score_infos <- lapply(clf_datas, getMarkerScoreInfo)
  all_score_info <- mergeScoreInfos(score_infos, verbose=T)

}
