% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{mergeScores}
\alias{mergeScores}
\title{Merge Scores}
\usage{
mergeScores(score.name, score.infos, aggr.func = c)
}
\arguments{
\item{score.name}{type of score to merge}

\item{score.infos}{scoring info per cell per cell type}
}
\description{
Merge Scores
}
