% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeSCExp.R
\name{normalizeSCExp}
\alias{normalizeSCExp}
\title{normalize single cell expression dataset}
\usage{
normalizeSCExp(expDat, total = 10000, dThresh = 0)
}
\arguments{
\item{expDat}{single cell expression dataset}

\item{total}{explanatory dataset}

\item{dThresh}{numeric or character variable indicating the response variable in response dataset}
}
\description{
normalize single cell expression dataset
}
\examples{

data(alldataset)
trainNorExp <- normalizeSCExp(alldataset$trainExplanatoryData,dThresh = 0.25)
testNorExp <- normalizeSCExp(alldataset$testExplanatoryData,dThresh = 0.25)

}
