% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{normalizeTfIdfWithFeatures}
\alias{normalizeTfIdfWithFeatures}
\title{Normalize TF-IDF with Features}
\usage{
normalizeTfIdfWithFeatures(cm, max.quantile = 0.95, max.smooth = 1e-10)
}
\arguments{
\item{cm}{matrix to normalize. Rows are observations (e.g. cells) and columns are features (e.g. genes)}

\item{max.quantile}{quantile to be used for max estimation during scaling}
}
\value{
Normalized matrix of the same shape as `cm`
}
\description{
Normalize `cm` matrix using TF-IDF and then column-wise min-max scaling
}
