% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markup_processing.R
\name{parseMarkerFile}
\alias{parseMarkerFile}
\title{Parse Marker File}
\usage{
parseMarkerFile(path, is.text = FALSE)
}
\arguments{
\item{path}{path to the markup file}
}
\value{
list of markers for each cell type. List elements have the following info: \itemize{
  \item{expressed: vector of positive markers}
  \item{not_expressed: vector of negative markers}
  \item{parent: parent cell type (`"root"` if there is no parent)}
}
}
\description{
read markers from the markup file
}
