% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plapply}
\alias{plapply}
\title{Parallel Lapply}
\usage{
plapply(..., n.cores = 1, verbose = F)
}
\arguments{
\item{n.cores}{number of cores to use}

\item{verbose}{show progress bar}
}
\description{
parallel, optionally verbose lapply
}
