% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotAssignmentScores}
\alias{plotAssignmentScores}
\title{Plot Assignment Scores}
\usage{
plotAssignmentScores(
  embedding,
  scores,
  classification.tree,
  parent.node = "root",
  build.panel = T,
  n.col = NULL,
  n.row = NULL,
  title.size = 12,
  ...
)
}
\arguments{
\item{scores}{assignment scores. Can be obtained with `assignCellsByScores`}

\item{classification.tree}{cell type hierarchy represented by graph. Part of `clf_data` object from `getClassificationData`}

\item{parent.node}{cell type, which subtypes' scores must be plotted}

\item{build.panel}{combine individual plots to a single panel}

\item{n.col}{number of columns in the panel}

\item{n.row}{number of rows in the panel}

\item{title.size}{font size for title of individual plots. Used only if build.panel==T}

\item{...}{
  Arguments passed on to \code{\link[conos:embeddingPlot]{conos::embeddingPlot}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
panel of plots if `build.panel==T` or `plot.list` otherwise
}
\description{
plot assignment scores on a separate scatterplot for each cell type
}
\examples{
  clf_data <- getClassificationData(cm, marker_path)
  ann_by_level <- assignCellsByScores(graph, clf_data)
  plotAssignmentScores(t_sne, ann_by_level$scores$l1, clf_data$classification.tree)
}
