% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAttr.R
\name{plotAttr}
\alias{plotAttr}
\title{heat map plot}
\usage{
plotAttr(predY, queryY)
}
\arguments{
\item{predY}{Predicted response variable}

\item{queryY}{A vector of response variable named with independent variables}
}
\description{
make heatmap plot between predicted and actual categories
}
\examples{

data(alldataset)
data(uniqueFeatureGene)
queryY <- alldataset$testResponseData$cellType
names(queryY) <- alldataset$testResponseData$cell
trainNorExp <- normalizeSCExp(alldataset$trainExplanatoryData,dThresh = 0.25)
x=t(as.matrix(trainNorExp[uniqueFeatureGene,]))
y=as.factor(alldataset$trainResponseData$cellType)
svmClassifier <- buildSVMclassifier(x,y)
testNorExp <- normalizeSCExp(alldataset$testExplanatoryData,dThresh = 0.25)
testX=t(as.matrix(testNorExp[uniqueFeatureGene,]))
predY <- predictWithclassifier(svmClassifier,testX)
\dontrun{
plotAttr(predY,queryY)
}

}
