% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotGeneExpression}
\alias{plotGeneExpression}
\title{Plot gene expression on cell embedding}
\usage{
plotGeneExpression(
  genes,
  embedding,
  cm,
  build.panel = T,
  n.col = NULL,
  n.row = NULL,
  title.size = 12,
  ...
)
}
\arguments{
\item{genes}{vector of gene names}

\item{embedding}{cell embedding}

\item{cm}{count matrix with genes as columns}

\item{build.panel}{combine individual plots to a single panel}

\item{n.col}{number of columns in the panel}

\item{n.row}{number of rows in the panel}

\item{title.size}{font size for title of individual plots. Used only if build.panel==T}

\item{...}{
  Arguments passed on to \code{\link[conos:embeddingPlot]{conos::embeddingPlot}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
Plot gene expression on cell embedding
}
