% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotMarkerListViolinMap}
\alias{plotMarkerListViolinMap}
\title{Plot expression violin map for given marker list}
\usage{
plotMarkerListViolinMap(
  count.matrix,
  annotation,
  parent.type = "root",
  clf.data = NULL,
  clf.tree = NULL,
  marker.list = NULL,
  max.depth = NULL,
  marker.type = "expressed",
  text.angle = 45,
  gene.order = NULL,
  ...
)
}
\arguments{
\item{count.matrix}{gene expression matrix with genes by columns and cells by rows}

\item{parent.type}{cell type for which the markers of the subtypes must be plotted}

\item{marker.list}{list of markers per cell type. Can be obtained with `parseMarkerFile`}
}
\description{
Plot expression violin map for given marker list
}
