% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotOneUncertaintyPerClust}
\alias{plotOneUncertaintyPerClust}
\title{Plot One Uncertainty per Cluster}
\usage{
plotOneUncertaintyPerClust(
  uncertainty.per.clust,
  clusters,
  annotation = NULL,
  ann.per.clust = NULL,
  threshold = 0.5,
  text.angle = 45,
  ...
)
}
\arguments{
\item{text.angle}{angle of x-axis labels}

\item{...}{
  Arguments passed on to \code{\link[sccore::styleEmbeddingPlot]{sccore:::styleEmbeddingPlot}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
Plot One Uncertainty per Cluster
}
