% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotUncertaintyPerCell}
\alias{plotUncertaintyPerCell}
\title{Uncertainty scatterplots per cell}
\usage{
plotUncertaintyPerCell(
  embedding,
  uncertainty.info,
  palette = colorRampPalette(c("gray", "#ffeda0", "#fec44f", "#f03b20")),
  alpha = 0.3,
  build.panel = T,
  n.col = length(uncertainty.info),
  n.row = NULL,
  title.size = 12,
  ...
)
}
\arguments{
\item{build.panel}{join plots to single panel}

\item{...}{
  Arguments passed on to \code{\link[conos:embeddingPlot]{conos::embeddingPlot}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
Uncertainty scatterplots per cell
}
