% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotUncertaintyPerClust}
\alias{plotUncertaintyPerClust}
\title{Uncertainty barplots per cluster}
\usage{
plotUncertaintyPerClust(uncertainty.info, annotation = NULL,
  clusters = NULL, ann.per.clust = NULL, thresholds = c(coverage =
  0.5, negative = 0.5, positive = 0.75), build.panel = T,
  n.col = length(uncertainty.info), n.row = NULL, ...)
}
\arguments{
\item{build.panel}{join plots to single panel}

\item{...}{Arguments passed on to \code{plotOneUncertaintyPerClust}
\describe{
  \item{text.angle}{angle of x-axis labels}
}}
}
\description{
Uncertainty barplots per cluster
}
