% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_selection.R
\name{prepareDeInfo}
\alias{prepareDeInfo}
\title{Prepare DE Info}
\usage{
prepareDeInfo(de.info, annotation, cm.raw, ..., n.cores = 1, verbose = FALSE)
}
\arguments{
\item{de.info}{list of data.frames with DE genes per cell type}

\item{annotation}{character vector with cell annotation, named with cell ids}

\item{cm.raw}{not normalized count matrix with cells by rows and genes by columns}
}
\description{
Prepare DE Info
}
