% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_expr.R
\docType{data}
\name{query_expr}
\alias{query_expr}
\title{Expression data of the query dataset}
\format{
A data frame with 2,000 rows and 200 variables
}
\usage{
data(query_expr)
}
\description{
A data set containing expression of 2,000 genes and 200 cells
}
\references{
{
Park J , Shrestha R , Qiu C , et al. Single-cell transcriptomics
of the mouse kidney reveals potential cellular targets of kidney
disease. Science, 2018, 360(6390):eaar2131.
}
}
\author{
Junhui Li \email{junhui.li11@umassmed.edu}
}
\keyword{data}
