% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_meta.R
\docType{data}
\name{query_meta}
\alias{query_meta}
\title{Meta data of the query data set}
\format{
A data frame with 200 rows and 2 variables
}
\usage{
data(query_meta)
}
\description{
A data set containing 200 cells and corresponding cell type.
The variable are as follows:
}
\details{
\itemize{
\item cell. the cell id
\item cellType. the corresponding cell type of each cell
}
}
\references{
{
Park J , Shrestha R , Qiu C , et al. Single-cell transcriptomics
of the mouse kidney reveals potential cellular targets of kidney
disease. Science, 2018, 360(6390):eaar2131.
}
}
\author{
Junhui Li \email{junhui.li11@umassmed.edu}
}
\keyword{data}
