% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_expr.R
\docType{data}
\name{reference_expr}
\alias{reference_expr}
\title{Expression data of the reference dataset}
\format{
A data frame with 2,000 rows and 200 variables
}
\usage{
data(reference_expr)
}
\description{
A data set containing expression of 2,000 genes and 200 cells
}
\references{
{
The Tabula Muris Consortium., Overall coordination., Logistical coordination. et al.
Single-cell transcriptomics of 20 mouse organs creates a Tabula Muris. Nature 562,
367–372 (2018). https://doi.org/10.1038/s41586-018-0590-4
}
}
\author{
Junhui Li \email{junhui.li11@umassmed.edu}
}
\keyword{data}
