% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_meta.R
\docType{data}
\name{reference_meta}
\alias{reference_meta}
\title{Meta data of the reference data set}
\format{
A data frame with 200 rows and 2 variables
}
\usage{
data(reference_meta)
}
\description{
A data set containing 200 cells and corresponding cell type.
The variable are as follows:
}
\details{
\itemize{
\item cell. the cell id
\item cellType. the corresponding cell type of each cell
}
}
\references{
{
The Tabula Muris Consortium., Overall coordination., Logistical coordination. et al.
Single-cell transcriptomics of 20 mouse organs creates a Tabula Muris. Nature 562,
367–372 (2018). https://doi.org/10.1038/s41586-018-0590-4
}
}
\author{
Junhui Li \email{junhui.li11@umassmed.edu}
}
\keyword{data}
