% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_estimation.R
\name{scoreCellUncertaintyPerLevel}
\alias{scoreCellUncertaintyPerLevel}
\title{Score Cell Uncertainty Per Level}
\usage{
scoreCellUncertaintyPerLevel(
  ann.info.per.level,
  score.info,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{ann.info.per.level}{annotation info per level. Result of `assignCellsByScores`.}

\item{score.info}{list of marker scores per cell type. Can be obtained with `getMarkerScoreInfo`}

\item{verbose}{show progress bar for estimation}

\item{...}{
  Arguments passed on to \code{\link[=scorePerCellUncertainty]{scorePerCellUncertainty}}
  \describe{
    \item{\code{cur.types}}{subset of types used to measure uncertainty. Default: all values presented in annotation}
    \item{\code{coverage.max.quantile}}{all coverage values above this quantille are winsorized}
  }}
}
\value{
Uncertainty per level for each cell
}
\description{
Estimate uncertainty scores per annotation level
}
\examples{
  score_info <- getMarkerScoreInfo(clf_data)
  unc_info <- scoreCellUncertaintyPerLevel(ann_by_level, score_info)

}
