% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_estimation.R
\name{scoreClusterUncertaintyPerLevel}
\alias{scoreClusterUncertaintyPerLevel}
\title{Score Cluster Uncertainty Per Level}
\usage{
scoreClusterUncertaintyPerLevel(cell.uncertainty.per.level, clusters)
}
\arguments{
\item{cell.uncertainty.per.level}{uncertainty per cell per level. Can be obtained from `scoreCellUncertaintyPerLevel`}

\item{clusters}{vector with cluster label for each cell}
}
\value{
Uncertainty per level for each cluster
}
\description{
Score uncertainty for each cluster per level
}
\examples{
  score_info <- getMarkerScoreInfo(clf_data)
  unc_info <- scoreCellUncertaintyPerLevel(ann_by_level, score_info)
  unc_per_clust <- scoreClusterUncertaintyPerLevel(unc_info, clusters)

}
