% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_estimation.R
\name{scorePerCellUncertainty}
\alias{scorePerCellUncertainty}
\title{Score Per Cell Uncertainty}
\usage{
scorePerCellUncertainty(
  annotation,
  scores.norm,
  score.info,
  cur.types = unique(annotation),
  coverage.max.quantile = 0.75
)
}
\arguments{
\item{annotation}{assigned annotation labels. Can be obtrained with `assignCellsByScores`.}

\item{scores.norm}{assignment scores. Can be obtrained with `assignCellsByScores`.}

\item{score.info}{list of marker scores per cell type. Can be obtained with `getMarkerScoreInfo`}

\item{cur.types}{subset of types used to measure uncertainty. Default: all values presented in annotation}

\item{coverage.max.quantile}{all coverage values above this quantille are winsorized}
}
\value{
list of scores with elements:\itemize{
  \item{positive: uncertainty based on conflicting positive markers}
  \item{negative: uncertainty based on large expression of negative markers}
  \item{coverage: uncertainty based on low coverage}
}
}
\description{
Score uncertainty per each cell
}
