% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_selection.R
\name{selectMarkersPerType}
\alias{selectMarkersPerType}
\title{Select Markers per Type}
\usage{
selectMarkersPerType(
  cm.norm,
  annotation,
  markers.per.type,
  marker.list = NULL,
  max.iters = nrow(cm.norm),
  parent = NULL,
  max.uncertainty = 0.25,
  verbose = 0,
  min.pos.markers = 1,
  max.pos.markers = 10,
  log.step = 1,
  n.cores = 1,
  refinement.period = 10,
  return.all = FALSE
)
}
\arguments{
\item{cm.norm}{tf-idf-normalized count matrix with cells by columns and genes by rows}

\item{markers.per.type}{marker candidate per type}

\item{marker.list}{list of already known markers per type, potentially empty}
}
\description{
Select Markers per Type
}
