% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitTrainTest.R
\name{splitTrainTest}
\alias{splitTrainTest}
\title{split dataset into train and test dataset}
\usage{
splitTrainTest(
  responseData,
  explanatoryData,
  groupBy,
  sampleID,
  splitParameter = c("ratio", "number"),
  splitRatio = 0.6,
  splitNum = 100,
  seed = 1234
)
}
\arguments{
\item{responseData}{response dataset}

\item{explanatoryData}{explanatory dataset}

\item{groupBy}{numeric or character variable indicating the response variable in response dataset}

\item{sampleID}{numeric or character variable indicating the sample id in response dataset}

\item{splitParameter}{which parameter to be used to split dataset}

\item{splitRatio}{ratio to split the dataset into train dataset}

\item{splitNum}{number of samples to split the dataset into train dataset, if the total sample number is less than splitNum, use splitRatio instead.}

\item{seed}{seed number for split}
}
\description{
split dataset into train and test dataset based on ratio of the number of each response variable or the given number
}
\examples{

data(reference_meta)
data(reference_expr)
datalist <- splitTrainTest(reference_meta,reference_expr,
groupBy="cellType",sampleID="cell",splitParameter="ratio")

}
