% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqueFeatureGene.R
\docType{data}
\name{uniqueFeatureGene}
\alias{uniqueFeatureGene}
\title{unique feature genes}
\format{
A vector with 40 variables
}
\usage{
data(uniqueFeatureGene)
}
\description{
A vector containing 40 genes.
The variable are as follows:
}
\author{
Junhui Li \email{junhui.li11@umassmed.edu}
}
\keyword{data}
