% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{assignCellsByScores}
\alias{assignCellsByScores}
\title{Assign cell types for each cell based on type scores. Optionally uses `clusters` to expand annotation.}
\usage{
assignCellsByScores(graph, clf.data, score.info = NULL,
  clusters = NULL, verbose = 0, ...)
}
\arguments{
\item{graph}{cell graph from Seurat, Pagoda2 or some other tool. Can be either in igraph or adjacency matrix format.
Use `graph=NULL` to skip graph diffusion step and get raw score annotation (useful when debug marker genes).}

\item{clf.data}{classification data from `getClassificationData`}

\item{score.info}{cell type scores from `getCellTypeScoreInfo` function. Re-estimated if NULL}

\item{clusters}{cluster assignment of data. Used to expand annotation on these clusters.}

\item{verbose}{verbosity level (from 0 to 2)}

\item{...}{Arguments passed on to \code{diffuseScorePerType}
\describe{
  \item{}{}
}}
}
\description{
Assign cell types for each cell based on type scores. Optionally uses `clusters` to expand annotation.
}
