% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{assignCellsByScores}
\alias{assignCellsByScores}
\title{Assign cell types for each cell based on type scores. Optionally uses `clusters` to expand annotation.}
\usage{
assignCellsByScores(graph, scores, classification.tree, clusters = NULL,
  ...)
}
\arguments{
\item{graph}{cell graph from Seurat, Pagoda2 or some other tool}

\item{scores}{cell type scores from `getMarkerScoresPerCellType` function}

\item{classification.tree}{cell type hierarchy from classification data (see `getClassificationData`)}

\item{clusters}{cluster assignment of data. Used to expand annotation on these clusters.}

\item{...}{Arguments passed on to \code{diffuseSubgraph}
\describe{
  \item{fading}{fading level for graph diffusion}
  \item{fading.const}{constant in exponent for graph diffusion}
  \item{verbose}{print progress bar}
  \item{tol}{tolerance for diffusion stopping}
}}
}
\description{
Assign cell types for each cell based on type scores. Optionally uses `clusters` to expand annotation.
}
