% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{diffuseGraph}
\alias{diffuseGraph}
\title{Run diffusion on graph}
\usage{
diffuseGraph(graph, scores, fading = 10, fading.const = 0.5,
  score.fixing.threshold = 0.8, clusters = NULL, verbose = FALSE,
  max.iters = 1000, tol = 0.001)
}
\arguments{
\item{fading}{fading level for graph diffusion}

\item{fading.const}{constant in exponent for graph diffusion}

\item{score.fixing.threshold}{threshold for a label to be considered true}

\item{verbose}{print progress bar}

\item{tol}{tolerance for diffusion stopping}
}
\description{
Run diffusion on graph
}
