% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getMarkerScoreInfo}
\alias{getMarkerScoreInfo}
\title{Get Marker Score Info}
\usage{
getMarkerScoreInfo(clf.data, ...)
}
\arguments{
\item{clf.data}{classification data from `getClassificationData`}
}
\description{
estimate info, neccessary for scoring of cells by cell types
}
