% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{normalizeTfIdfWithFeatures}
\alias{normalizeTfIdfWithFeatures}
\title{Normalize TF-IDF with Features}
\usage{
normalizeTfIdfWithFeatures(cm, ...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=normalizeTCWithFeatures]{normalizeTCWithFeatures}}
  \describe{
    \item{\code{max.quantile}}{quantile to be used for max estimation during scaling}
  }}
}
\description{
Normalize `cm` matrix using total counts, column-wise min-max scaling and then IDF normalization
}
