% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotMarkerListViolinMap}
\alias{plotMarkerListViolinMap}
\title{Plot expression violin map for given marker list}
\usage{
plotMarkerListViolinMap(count.matrix, annotation, parent.type = "root",
  clf.data = NULL, clf.tree = NULL, marker.list = NULL,
  max.depth = NULL, marker.type = "expressed", text.angle = 45,
  gene.order = NULL, ...)
}
\description{
Plot expression violin map for given marker list
}
