% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotSubtypeMarkers}
\alias{plotSubtypeMarkers}
\title{Plot Subtype Markers}
\usage{
plotSubtypeMarkers(
  embedding,
  count.matrix,
  parent.type = "root",
  clf.data = NULL,
  clf.tree = NULL,
  marker.list = NULL,
  show.legend = F,
  max.depth = NULL,
  build.panel = T,
  n.col = NULL,
  n.row = NULL,
  marker.type = c("expressed", "not_expressed"),
  title.size = 10,
  ...
)
}
\arguments{
\item{count.matrix}{gene expression matrix with genes by columns and cells by rows}

\item{parent.type}{cell type for which the markers of the subtypes must be plotted}

\item{marker.list}{list of markers per cell type. Can be obtained with `parseMarkerFile`}

\item{build.panel}{combine individual plots to a single panel}

\item{n.col}{number of columns in the panel}

\item{n.row}{number of rows in the panel}

\item{title.size}{font size for title of individual plots. Used only if build.panel==T}

\item{...}{
  Arguments passed on to \code{\link[conos:embeddingPlot]{conos::embeddingPlot}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
plot markers for all subtypes of the specified `parent.type`
}
