% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotUncertaintyPerClust}
\alias{plotUncertaintyPerClust}
\title{Uncertainty barplots per cluster}
\usage{
plotUncertaintyPerClust(uncertainty.per.clust, clusters,
  annotation = NULL, ann.per.clust = NULL, thresholds = c(coverage =
  0.5, negative = 0.5, positive = 0.75), build.panel = F, n.col = 3,
  n.row = NULL, ...)
}
\arguments{
\item{...}{Arguments passed on to \code{plotOneUncertaintyPerClust}
\describe{
  \item{text.angle}{angle of x-axis labels}
}}
}
\description{
Uncertainty barplots per cluster
}
