% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotUncertaintyPerClust}
\alias{plotUncertaintyPerClust}
\title{Uncertainty barplots per cluster}
\usage{
plotUncertaintyPerClust(
  uncertainty.per.clust,
  clusters,
  annotation = NULL,
  ann.per.clust = NULL,
  thresholds = c(coverage = 0.5, negative = 0.5, positive = 0.75),
  build.panel = TRUE,
  n.col = 1,
  n.row = NULL,
  adjust.legend = TRUE,
  rel.legend.width = 0.25,
  title.size = 12,
  ...
)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=plotOneUncertaintyPerClust]{plotOneUncertaintyPerClust}}
  \describe{
    \item{\code{text.angle}}{angle of x-axis labels}
  }}
}
\description{
Uncertainty barplots per cluster
}
